requirements.txt
Flask==3.0.0
google-genai>=0.2.0
requests==2.31.0
beautifulsoup4==4.12.0
SQLAlchemy==2.0.43
typing-extensions>=4.6.0
python-dotenv==1.0.0
newsapi-python==0.2.7
google-api-python-client==2.108.0
Flask-SQLAlchemy==3.1.1
urllib3==2.1.0
